refactor: Gnss and Sbas List into it's own file
